requirements.txt
flask
flask-cors
flask-restful
psycopg2-binary
python-dotenv